v2/requirements.txt
flask
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-pubsub
requests
beautifulsoup4
selenium
numpy
pandas